backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1

# HTTP requests (Shopify API)
requests==2.32.3

# Data processing
pandas==2.2.2
numpy>=1.26.0

# AI - OpenAI
openai==1.35.0
httpx==0.27.2

# Environment variables
python-dotenv==1.0.1
